requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.13.1,<3.14
psycopg2>=2.8.6,<2.9
psycopg2-binary<2.9,>=2.8.6
drf-spectacular>=0.15.1,<0.16
django-crispy-forms>=1.13.0
uwsgi>=2.0.20,<2.1
Pillow==9.3.0
six==1.16.0
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-crispy-forms
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-widget-tweaks
requests
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
python-dotenv==0.20.0
# requests==2.28.1
elasticsearch===7.17.7
elasticsearch-dsl==7.4.1
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
urllib3>=1.21.1,<2
django-filter
Faker==4.1.1
python-dateutil==2.8.1
asgiref
pytz==2020.1
sqlparse==0.3.1
text-unidecode==1.3
